+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - Add Sphinx documentation infrastructure and skeleton
 - Fix search metadata reset behavior
 - Activate IssueReporter email target again
+- Update IFI Claims documentation
 
 2017-10-13 0.141.0
 ==================