+++ CHANGES.rst
@@ -8,6 +8,7 @@ development
 - Fix search metadata reset behavior
 - Activate IssueReporter email target again
 - Update IFI Claims documentation
+- Fix dependencies in setup.py for compatibility between Mac OSX and Debian GNU/Linux
 
 2017-10-13 0.141.0
 ==================